fixes string vals in int map styles, +tests

+++ models/NeatlineDataRecord.php
@@ -167,16 +167,16 @@ class NeatlineDataRecord extends Omeka_record
         $data['vector_color'] =     $this->getStyle('vector_color');
         $data['stroke_color'] =     $this->getStyle('stroke_color');
         $data['highlight_color'] =  $this->getStyle('highlight_color');
-        $data['vector_opacity'] =   $this->getStyle('vector_opacity');
-        $data['stroke_opacity'] =   $this->getStyle('stroke_opacity');
-        $data['stroke_width'] =     $this->getStyle('stroke_width');
-        $data['point_radius'] =     $this->getStyle('point_radius');
+        $data['vector_opacity'] =   (int) $this->getStyle('vector_opacity');
+        $data['stroke_opacity'] =   (int) $this->getStyle('stroke_opacity');
+        $data['stroke_width'] =     (int) $this->getStyle('stroke_width');
+        $data['point_radius'] =     (int) $this->getStyle('point_radius');
         $data['start_date'] =       (string) $this->start_date;
         $data['start_time'] =       (string) $this->start_time;
         $data['end_date'] =         (string) $this->end_date;
         $data['end_time'] =         (string) $this->end_time;
-        $data['left_percent'] =     $this->getLeftPercent();
-        $data['right_percent'] =    $this->getRightPercent();
+        $data['left_percent'] =     (int) $this->getLeftPercent();
+        $data['right_percent'] =    (int) $this->getRightPercent();
 
         // JSON-ify the array.
         return json_encode($data);
@@ -200,10 +200,10 @@ class NeatlineDataRecord extends Omeka_record
         $data['vector_color'] =     get_option('vector_color');
         $data['stroke_color'] =     get_option('stroke_color');
         $data['highlight_color'] =  get_option('highlight_color');
-        $data['vector_opacity'] =   get_option('vector_opacity');
-        $data['stroke_opacity'] =   get_option('stroke_opacity');
-        $data['stroke_width'] =     get_option('stroke_width');
-        $data['point_radius'] =     get_option('point_radius');
+        $data['vector_opacity'] =   (int) get_option('vector_opacity');
+        $data['stroke_opacity'] =   (int) get_option('stroke_opacity');
+        $data['stroke_width'] =     (int) get_option('stroke_width');
+        $data['point_radius'] =     (int) get_option('point_radius');
         $data['left_percent'] =     self::$defaults['left_percent'];
         $data['right_percent'] =    self::$defaults['right_percent'];
         $data['start_date'] =       '';
